+++ src/index.ts
@@ -32,7 +32,10 @@ export { default as DatePicker } from "./components/DatePicker/DatePicker";
 export { DatePickerTypes } from "./components/DatePicker/DatePickerTypes";
 export { default as Form, FormField, FormRow } from "./components/Form/Form";
 export { default as Heading } from "./components/Heading/Heading";
-export { HeadingDisplaySizes } from "./components/Heading/HeadingDisplaySizes";
+export {
+  HeadingDisplaySizes,
+  HeadingLevels,
+} from "./components/Heading/HeadingTypes";
 export { default as HelperErrorText } from "./components/HelperErrorText/HelperErrorText";
 export { default as Hero } from "./components/Hero/Hero";
 export { HeroTypes } from "./components/Hero/HeroTypes";